refactor(slang): share address_type via expression/mod.rs and short-circuit sol.load
Moves `address_type` from `access.rs` into `ExpressionEmitter` so future
gep/map callers (struct fields, etc.) share the same non-ptr-ref-in-storage
rule. Adds a same-type short-circuit at the top of `Builder::emit_sol_load`,
mirroring `emit_sol_cast`, so the read path in `access.rs::emit_index_access`
no longer needs its own `if address.r#type() == element_type` branch — it
just calls `emit_sol_load` unconditionally. The `anyhow::bail!` arms in
`emit_assignment` switch to `unreachable!` (slang resolves identifiers) and
`unimplemented!` (codegen gap), matching the project rule that reserves
`bail!` for genuine resolution errors.
