+++ solx-slang/src/ast/contract/function/expression/access.rs
@@ -5,12 +5,10 @@
 use melior::ir::BlockRef;
 use melior::ir::Type;
 use melior::ir::Value;
-use melior::ir::ValueLike;
 use slang_solidity::backend::ir::ast::IndexAccessExpression;
 use slang_solidity::backend::ir::ast::Type as SlangType;
 use slang_solidity::backend::types::DataLocation as SlangDataLocation;
 
-use solx_mlir::Builder;
 use solx_utils::DataLocation;
 
 use crate::ast::contract::function::expression::ExpressionEmitter;
@@ -26,13 +24,10 @@ impl<'state, 'context, 'block> ExpressionEmitter<'state, 'context, 'block> {
         block: BlockRef<'context, 'block>,
     ) -> anyhow::Result<(Option<Value<'context, 'block>>, BlockRef<'context, 'block>)> {
         let (address, element_type, block) = self.emit_index_access_address(index_access, block)?;
-        let value = if address.r#type() == element_type {
-            address
-        } else {
-            self.state
-                .builder
-                .emit_sol_load(address, element_type, &block)?
-        };
+        let value = self
+            .state
+            .builder
+            .emit_sol_load(address, element_type, &block)?;
         Ok((Some(value), block))
     }
 
@@ -105,30 +100,6 @@ impl<'state, 'context, 'block> ExpressionEmitter<'state, 'context, 'block> {
                 "index access on a value-typed base is not yet wired: {:?}",
                 std::mem::discriminant(base_slang_type)
             ),
-        }
-    }
-
-    /// Picks the MLIR type of the address yielded by `sol.gep` / `sol.map`.
-    ///
-    /// Mirrors `Sol_GepOp::build`'s non-ptr-ref-in-storage rule: when the
-    /// element is itself a reference type and lives in `Storage` or
-    /// `CallData`, the result address IS the element type rather than a
-    /// pointer to it.
-    fn address_type(
-        builder: &Builder<'context>,
-        element_type: Type<'context>,
-        base_location: DataLocation,
-        result_slang_type: &SlangType,
-    ) -> Type<'context> {
-        if result_slang_type.is_reference_type()
-            && matches!(
-                base_location,
-                DataLocation::Storage | DataLocation::CallData
-            )
-        {
-            element_type
-        } else {
-            builder.types.pointer(element_type, base_location)
         }
     }
 }
